pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "asyncblobdict"
version = "0.1.12"
description = "Minimal async key-value store backed by local or cloud blob storage"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "vilahte", email = "vilahte@gmail.com"}
]
dependencies = [
    "azure-storage-blob",
    "asyncio",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Database",
]
keywords = ["blob storage", "adapter", "async", "key-value store", "azure", "local filesystem"]


[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "python-dotenv"
]

